requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
scikit-learn
feedparser
numpy
pandas
requests
json

